+++ xbmc/cores/dvdplayer/Edl.h
@@ -40,8 +40,8 @@ public:
 
   struct Cut
   {
-    int64_t start; // ms
-    int64_t end;   // ms
+    int start; // ms
+    int end;   // ms
     Action action;
   };
 
@@ -51,23 +51,23 @@ public:
   bool HasCut();
   bool HasSceneMarker();
   std::string GetInfo();
-  int64_t GetTotalCutTime();
-  int64_t RemoveCutTime(int64_t iSeek);
-  int64_t RestoreCutTime(int64_t iClock);
+  int GetTotalCutTime();
+  int RemoveCutTime(int iSeek);
+  int RestoreCutTime(int iClock);
 
-  bool InCut(int64_t iSeek, Cut *pCut = NULL);
+  bool InCut(int iSeek, Cut *pCut = NULL);
 
-  bool GetNextSceneMarker(bool bPlus, const int64_t iClock, int64_t *iSceneMarker);
+  bool GetNextSceneMarker(bool bPlus, const int iClock, int *iSceneMarker);
 
   static std::string GetMPlayerEdl();
 
-  static std::string MillisecondsToTimeString(const int64_t iMilliseconds);
+  static std::string MillisecondsToTimeString(const int iMilliseconds);
 
 protected:
 private:
-  int64_t m_iTotalCutTime; // ms
+  int m_iTotalCutTime; // ms
   std::vector<Cut> m_vecCuts;
-  std::vector<int64_t> m_vecSceneMarkers;
+  std::vector<int> m_vecSceneMarkers;
 
   bool ReadEdl(const std::string& strMovie, const float fFramesPerSecond);
   bool ReadComskip(const std::string& strMovie, const float fFramesPerSecond);
@@ -78,7 +78,7 @@ private:
   bool ReadMythCutList(const std::string& strMovie, const float fFramesPerSecond);
 
   bool AddCut(Cut& NewCut);
-  bool AddSceneMarker(const int64_t sceneMarker);
+  bool AddSceneMarker(const int sceneMarker);
 
   bool WriteMPlayerEdl();
